refactor: consolidate calculator implementations - Removed separate Ruby namespace implementation - Consolidated batch and single calculation logic into one file - Moved position calculation into PositionCalculator class - Improved code organization and reduced complexity - Fixed RuboCop issues in the consolidated implementation
